+++ devel/Dual endpoint separate model 27012015.R
@@ -1,3 +1,5 @@
+# nolint start
+
 library(crmPack)
 mySize<-CohortSizeConst(size=3)
 myIncrements1 <-IncrementsRelative(intervals=c(2),increments=c(2))
@@ -44,8 +46,7 @@ doseRecGain
 
 
 #stopping
-myStopping7 <- StoppingGstarCIRatio(targetRatio=5, 
-                                    targetEndOfTrial=0.3)
+myStopping7 <- StoppingMaxGainCIRatio(target_ratio = 5, prob_target = 0.3)
 
 myStopping8 <- myStopping7 | StoppingMinPatients(72)
 stopTrial(stopping=myStopping7,dose=doseRecGain$nextdose,model=newDLTmodel,
@@ -68,9 +69,11 @@ myfun <- function(x, a, b)
 }
 test <- function(x, a, b)
 {
-lapply(x, 
+lapply(x,
        myfun,
        a,
        b)
 }
 test(1:5, a=3, b=2)
+
+# nolint end